[feature #165412917] rename article controller
