+++ src/tests/controllers/index.js
@@ -1,3 +1,3 @@
 import './CommentController.test';
 import './AuthPassportController.test';
-import './Articles.test';
+import './ArticleContoller.test';
